Alterando logica do cadastro de postagem e seus testes

+++ src/main/java/rede/social/nester/services/PostagemService.java
@@ -8,12 +8,14 @@ import org.springframework.stereotype.Service;
 import jakarta.transaction.Transactional;
 import rede.social.nester.entities.PostagemEntity;
 import rede.social.nester.entities.UsuarioEntity;
+import rede.social.nester.enuns.UsuarioEnum;
 import rede.social.nester.exceptions.NotFoundBussinessException;
+import rede.social.nester.exceptions.UnauthorizedAccessBussinessException;
 import rede.social.nester.repositories.PostagemRepository;
 
 @Service
 public class PostagemService {
-	
+
 	@Autowired
 	private PostagemRepository postagemRepository;
 
@@ -22,22 +24,30 @@ public class PostagemService {
 		return postagemRepository.save(postagemEntity);
 	}
 
-
-    public List<PostagemEntity> buscaPostagemDoUsuario(UsuarioEntity usuarioEncontrado) {
+	public List<PostagemEntity> buscaPostagemDoUsuario(UsuarioEntity usuarioEncontrado) {
 		return postagemRepository.findAllByUsuario(usuarioEncontrado);
-    }
+	}
 
 	public PostagemEntity buscaPostagemPeloId(Long id) {
 		return postagemRepository.findById(id)
 				.orElseThrow(() -> new NotFoundBussinessException("Postagem não encontrada"));
 	}
 
-	public void removerPostagem(PostagemEntity postagem) {
-		postagemRepository.delete(postagem);
+	@Transactional
+	public void removerPostagem(UsuarioEntity usuarioEncontrado, PostagemEntity postagem) {
+		if (usuarioEncontrado.getRole() == UsuarioEnum.ADMIN || usuarioEncontrado == postagem.getUsuario()) {
+			postagemRepository.delete(postagem);
+		} else {
+			throw new UnauthorizedAccessBussinessException("Usuario não tem permissão necessária!");
+		}
 	}
 
 	@Transactional
-	public PostagemEntity atualizaPostagem(PostagemEntity postagemEncontrada) {
-		return postagemRepository.save(postagemEncontrada);
+	public PostagemEntity atualizarPostagem(UsuarioEntity usuarioEncontrado, PostagemEntity postagem) {
+		if (usuarioEncontrado.getRole() == UsuarioEnum.ADMIN || usuarioEncontrado == postagem.getUsuario()) {
+			return postagemRepository.save(postagem);
+		} else {
+			throw new UnauthorizedAccessBussinessException("Usuario não tem permissão necessária!");
+		}
 	}
 }
